+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 1.7 (unreleased)
 ----------------
 
-- ``simplejson`` not required on Python 2.7+
+- Removed deprecated ``simplejson`` dependency
   [keul]
 
 - Removed plone.app.jquerytools from generic setup dependencies.